API struct to represent the firmware manifest
This is used as the leaf in the binary transparency log. This is the other side of PR #69.

Note there is something a bit funny going on here that we need to get to the bottom of. I've advertised both of the digests as being sha256 but that doesn't seem to be true. A sha256 hash should have 32 bytes, but the example pulled from #69 has a CommitHash of length 30 and a Sha256Digest of length 48.