Don't embed build info (#180)
This information isn't useful and makes it hard to reproducibly build the same output. If it was still useful then we could add it to the manifest, but it isn't, and so it is being removed. The build epcoh that is merged into CI builds will also need consideration, but is a separate PR.