Bump CI version to 0.3.X (#205)
This was downgraded from 0.2.X to 0.0.X because it seemed arbitrary. This caused a latent problem that has now come to light as we're trying to flash these CI builds onto the devices, which is that recent builds appear older than those from before the version downgrade. In the interests of being pragmatic, bumping this version to a newer version bypasses this problem. We can reset this to 0.0.X again should we ever rotate the CI logs.