Commit to all build envs in manifest (#227)
Any of these values could be rotated in the future, and given they are critical for reproducible builds they should be in the manifest. One could argue that these should have their own top-level named entry in the manifest rather than being buried in build_envs. For pragmatism, let's use this approach for now and we can revisit this in a month as part of a whole manifest refactor before we commit to v1.