Cache DNS responses for a minute (#257)
Previously this was making a DNS query for every request. Given that we have a very small number of DNS lookups that we make, and we git them relatively frequently, a cache cuts down on the load we put on the DNS server, and the latency involved in waiting for it before each request.